requirements.txt
flask>=3.0
werkzeug>=3.0
python-telegram-bot>=21.0
gunicorn>=22.0
